% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionplot.R
\name{functionplotOutput}
\alias{functionplotOutput}
\title{Widget output function for use in Shiny}
\usage{
functionplotOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

