% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ functionplotR }
\alias{ functionplotR }
\alias{ functionplotR }
\title{ function-plot d3.js library as an R htmlwidget }
\description{
function-plot d3.js library as an R htmlwidget
}
