% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionplot.R
\name{renderFunctionplot}
\alias{renderFunctionplot}
\title{Widget render function for use in Shiny}
\usage{
renderFunctionplot(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

